requirements.txt
django==3.2
django-ninja==0.13.0
djangorestframework==3.12.4
pandas==1.2.4
